+++ scripts/const.js
@@ -1,0 +1,7 @@
+/* globals
+
+*/
+"use strict";
+
+export const MODULE_ID = "elevationruler";
+export const EPSILON = 1e-08;